refactored out multi-class files so each class had their own file, conforming to PSR-4
