+++ src/Elo/FideKFactor.php
@@ -15,14 +15,4 @@ class FideKFactor extends KFactor
     }    
 }
 
-/**
- * Indicates someone who has played less than 30 games.
- */
-class ProvisionalFideKFactor extends FideKFactor
-{
-    public function getValueForRating($rating)
-    {
-        return 25;
-    }
-}
 ?>
