+++ src/FactorGraphs/Variable.php
@@ -36,38 +36,4 @@ class Variable
     }
 }
 
-class DefaultVariable extends Variable
-{
-    public function __construct()
-    {
-        parent::__construct("Default", null);
-    }
-
-    public function getValue()
-    {
-        return null;
-    }
-
-    public function setValue($value)
-    {
-        throw new Exception();
-    }
-}
-
-class KeyedVariable extends Variable
-{
-    private $_key;
-    public function __construct($key, $name, $prior)
-    {
-        parent::__construct($name, $prior);
-        $this->_key = $key;
-    }
-
-    public function getKey()
-    {
-        $key = $this->_key;
-        return $key;
-    }
-}
-
 ?>
