+++ src/InvalidOperationException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Skills;
+
+class InvalidOperationException extends \Exception {
+
+}
