+++ src/Numerics/IdentityMatrix.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Skills\Numerics;
+
+class IdentityMatrix extends DiagonalMatrix
+{
+  public function __construct($rows)
+  {
+    parent::__construct(\array_fill(0, $rows, 1));
+  }
+}
