+++ src/Numerics/Vector.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Skills\Numerics;
+
+class Vector extends Matrix
+{
+  public function __construct(array $vectorValues)
+  {
+    $columnValues = array();
+    foreach($vectorValues as $currentVectorValue)
+    {
+      $columnValues[] = array($currentVectorValue);
+    }
+    parent::__construct(count($vectorValues), 1, $columnValues);
+  }
+}
