+++ src/SkillCalculatorSupportedOptions.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: les.peabody
+ * Date: 10/13/15
+ * Time: 11:13 PM
+ */
+
+namespace Skills;
+
+class SkillCalculatorSupportedOptions
+{
+  const NONE = 0x00;
+  const PARTIAL_PLAY = 0x01;
+  const PARTIAL_UPDATE = 0x02;
+}
