+++ src/TrueSkill/Layers/PlayerSkillsToPerformancesLayer.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace Skills\TrueSkill\Layers;
 
 use Skills\FactorGraphs\ScheduleStep;
