+++ src/TrueSkill/Layers/TrueSkillFactorGraphLayer.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace Skills\TrueSkill\Layers;
 
 use Skills\FactorGraphs\FactorGraphLayer;
