+++ tests/Numerics/BasicMathTest.php
@@ -1,13 +1,16 @@
 <?php
 namespace Skills\Tests\Numerics;
 
+use Skills\Numerics\BasicMath;
+use PHPUnit_Framework_TestCase;
+
 class BasicMathTest extends PHPUnit_Framework_TestCase
 {    
     public function testSquare()
     {    
-        $this->assertEquals( 1, Moserware\Numerics\square(1) );
-        $this->assertEquals( 1.44, Moserware\Numerics\square(1.2) );
-        $this->assertEquals( 4, Moserware\Numerics\square(2) );
+        $this->assertEquals( 1, BasicMath::square(1) );
+        $this->assertEquals( 1.44, BasicMath::square(1.2) );
+        $this->assertEquals( 4, BasicMath::square(2) );
     }
 }
 ?>
