+++ tests/Numerics/GaussianDistributionTest.php
@@ -1,6 +1,10 @@
 <?php
 namespace Skills\Tests\TrueSkill;
 
+use PHPUnit_Framework_TestCase;
+use Skills\Numerics\GaussianDistribution;
+use Skills\Numerics\BasicMath;
+
 class GaussianDistributionTest extends PHPUnit_Framework_TestCase
 {    
     const ERROR_TOLERANCE = 0.000001;
@@ -34,10 +38,10 @@ class GaussianDistributionTest extends PHPUnit_Framework_TestCase
 
         $product2 = GaussianDistribution::multiply($m4s5, $m6s7);
         
-        $expectedMean = (4 * square(7) + 6 * square(5)) / (square(5) + square(7));
+        $expectedMean = (4 * BasicMath::square(7) + 6 * BasicMath::square(5)) / (BasicMath::square(5) + BasicMath::square(7));
         $this->assertEquals($expectedMean, $product2->getMean(), '', GaussianDistributionTest::ERROR_TOLERANCE);
 
-        $expectedSigma = sqrt(((square(5) * square(7)) / (square(5) + square(7))));
+        $expectedSigma = sqrt(((BasicMath::square(5) * BasicMath::square(7)) / (BasicMath::square(5) + BasicMath::square(7))));
         $this->assertEquals($expectedSigma, $product2->getStandardDeviation(), '', GaussianDistributionTest::ERROR_TOLERANCE);
     }
     
@@ -51,7 +55,7 @@ class GaussianDistributionTest extends PHPUnit_Framework_TestCase
         $this->assertEquals(2.0, $productDividedByStandardNormal->getMean(), '', GaussianDistributionTest::ERROR_TOLERANCE);
         $this->assertEquals(3.0, $productDividedByStandardNormal->getStandardDeviation(),'', GaussianDistributionTest::ERROR_TOLERANCE);
         
-        $product2 = new GaussianDistribution((4 * square(7) + 6 * square(5)) / (square(5) + square(7)), sqrt(((square(5) * square(7)) / (square(5) + square(7)))));
+        $product2 = new GaussianDistribution((4 * BasicMath::square(7) + 6 * BasicMath::square(5)) / (BasicMath::square(5) + BasicMath::square(7)), sqrt(((BasicMath::square(5) * BasicMath::square(7)) / (BasicMath::square(5) + BasicMath::square(7)))));
         $m4s5 = new GaussianDistribution(4,5);
         $product2DividedByM4S5 = GaussianDistribution::divide($product2, $m4s5);
         $this->assertEquals(6.0, $product2DividedByM4S5->getMean(), '', GaussianDistributionTest::ERROR_TOLERANCE);
