+++ tests/TrueSkill/TwoTeamTrueSkillCalculatorTest.php
@@ -2,9 +2,7 @@
 namespace Skills\Tests\TrueSkill;
 
 use Skills\TrueSkill\TwoTeamTrueSkillCalculator;
-use Skills\Tests\TrueSkill\TrueSkillCalculatorTests;
 use PHPUnit_Framework_TestCase;
-use PHPUnit_TextUI_TestRunner;
 
 class TwoTeamTrueSkillCalculatorTest extends PHPUnit_Framework_TestCase
 {
@@ -13,13 +11,10 @@ class TwoTeamTrueSkillCalculatorTest extends PHPUnit_Framework_TestCase
         $calculator = new TwoTeamTrueSkillCalculator();
 
         // We only support two players
+        //TODO: uncomment testAllTwoPlayerScenarios
         TrueSkillCalculatorTests::testAllTwoPlayerScenarios($this, $calculator);
         TrueSkillCalculatorTests::testAllTwoTeamScenarios($this, $calculator);
     }
 }
 
-$testSuite = new \PHPUnit_Framework_TestSuite();
-$testSuite->addTest( new TwoTeamTrueSkillCalculatorTest("testTwoTeamTrueSkillCalculator"));
-
-PHPUnit_TextUI_TestRunner::run($testSuite);
 ?>
